+++ zcode/math/tests/test_math_core.py
@@ -4,6 +4,8 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import numpy as np
 from numpy.testing import run_module_suite
+import scipy as sp
+import scipy.stats
 from nose.tools import assert_true, assert_false, assert_equal, assert_raises
 
 from zcode.math import math_core
@@ -38,6 +40,20 @@ class TestMathCore(object):
         assert_true(math_core.mono(arr_e, 'e'))
         assert_false(math_core.mono(arr_le, 'e'))
 
+    def test_really1d(self):
+        from zcode.math import really1d
+        assert_true(really1d([1, 2, 3]))
+        assert_true(really1d([1]))
+        assert_true(really1d([]))
+        assert_true(really1d(np.arange(10)))
+
+        assert_false(really1d(1))
+        assert_false(really1d([[1]]))
+        assert_false(really1d([[1, 2], [2, 3]]))
+        assert_false(really1d([[1, 2, 3], [4, 5]]))
+        assert_false(really1d(np.random.random((4, 3))))
+        assert_false(really1d([[]]))
+
     def test_argextrema(self):
         # Basic usage without filtering
         assert_equal(math_core.argextrema([-1, -5, 2, 10], 'min'), 1)
@@ -65,6 +81,110 @@ class TestMathCore(object):
         assert_raises(ValueError, math_core.argextrema, [1, 2], 'max', 'e')
         #    Invalid `filter` argument
         assert_raises(ValueError, math_core.argextrema, [1, 2], 'max', 'greater')
+
+    def test_asBinEdges_1d(self):
+        print("TestMathCore.test_asBinEdges_1d")
+        from zcode.math import asBinEdges, spacing
+
+        data_1d = np.random.random(40)
+        bins_1d = np.arange(20)
+        # Preserves valid bins
+        assert_true(np.allclose(bins_1d, asBinEdges(bins_1d, data_1d)))
+
+        # Constructs valid bins
+        #    lin
+        lin_1d = spacing(data_1d, scale='lin', num=8+1)
+        lin_edges_1d = asBinEdges(8, data_1d, scale='lin')
+        assert_true(np.allclose(lin_1d, lin_edges_1d))
+        #    log
+        log_1d = spacing(data_1d, scale='log', num=7+1)
+        log_edges_1d = asBinEdges(7, data_1d, scale='log')
+        assert_true(np.allclose(log_1d, log_edges_1d))
+
+        # Raises appropriate errors
+        data_2d = data_1d.reshape(8, 5)
+        bins_2d = bins_1d.reshape(4, 5)
+        #    1D bins, 2D data
+        assert_raises(ValueError, asBinEdges, bins_1d, data_2d)
+        #    2D bins, 1D data
+        assert_raises(ValueError, asBinEdges, bins_2d, data_1d)
+
+    def test_asBinEdges_nd(self):
+        print("TestMathCore.test_asBinEdges_nd")
+        from zcode.math import asBinEdges
+
+        data_2d = np.random.random((8, 2))
+        bins_2d = np.arange(8).reshape(2, 4)
+        bins_2d2 = [[0.0, 1.0], [0.0, 0.5, 1.0]]
+
+        # Preserves valid bins
+        edges_2d = asBinEdges(bins_2d, data_2d)
+        assert_true(np.allclose(bins_2d, edges_2d))
+        edges_2d2 = asBinEdges(bins_2d2, data_2d)
+        assert_true(np.allclose(bins_2d2[0], edges_2d2[0]))
+        assert_true(np.allclose(bins_2d2[1], edges_2d2[1]))
+
+        # Constructs valid bins
+        #    lin
+        lin_2d1 = sp.stats.binned_statistic_dd(data_2d, None, 'count', bins=4).bin_edges
+        lin_edges_2d1 = asBinEdges(4, data_2d, scale='lin')
+        assert_true(np.allclose(lin_2d1, lin_edges_2d1))
+        lin_2d2 = sp.stats.binned_statistic_dd(data_2d, None, 'count', bins=[4, 3]).bin_edges
+        lin_edges_2d2 = asBinEdges([4, 3], data_2d, scale='lin')
+        assert_true(np.allclose(lin_2d2[0], lin_edges_2d2[0]))
+        assert_true(np.allclose(lin_2d2[1], lin_edges_2d2[1]))
+
+        # Raises appropriate errors
+        #    1D bins, 2D data
+        assert_raises(ValueError, asBinEdges, [4], data_2d)
+        #    2D bins, 1D data
+        assert_raises(ValueError, asBinEdges, [4, 3, 2], data_2d)
+
+    def test_confidenceBands(self):
+        print("TestMathCore.test_confidenceBands")
+        from zcode.math import confidenceBands
+        np.random.seed(9865)
+        uni1 = np.random.uniform(0.0, 1.0, size=100)
+        uni2 = np.random.uniform(0.0, 1.0, size=100)
+        zz = np.linspace(-np.pi, np.pi, num=20)
+        yy = np.array([uni1 + np.sin(z) for z in zz])
+        dz = np.average(np.diff(zz))
+        xx = np.array([4.0*dz*(uni2-0.5) + z for z in zz])
+
+        count, med, conf, xbins = confidenceBands(xx, yy, 20, 'lin', confInt=0.68)
+        true_count = np.array([43,  68, 100, 115, 111, 114, 118, 121, 113, 110,
+                               117, 112, 114, 120, 115, 114, 111,  95,  58,  31])
+        true_med = np.array([0.44741627,  0.27070286,  0.13062732, -0.06400147, -0.30522202,
+                             -0.35270415, -0.32772738, -0.21848981,  0.0700967,  0.38662969,
+                             0.75191171,  1.07342266,  1.31406718,  1.43435843,  1.46189632,
+                             1.31720551,  1.13144681,  0.92959401,  0.78686968,  0.67194954])
+        true_conf = [[0.12095047, -0.17810892, -0.38115803, -0.51990728, -0.66618246,
+                      -0.70931296, -0.6526213, -0.56719675, -0.30787827, -0.0927599,
+                      0.22306699,  0.59704074,  0.93804672,  1.11844012,  1.05683467,
+                      0.81939978,  0.57008734,  0.48195629,  0.35877874,  0.37992886],
+                     [0.836671,  0.6102549,  0.48461273,  0.378781,  0.06250324,
+                      -0.07095468, -0.04947349,  0.12198184,  0.53509063,  0.87365337,
+                      1.18369298,  1.45002858,  1.62060513,  1.71902407,  1.76211033,
+                      1.63896559,  1.49360737,  1.39381561,  1.17244189,  0.9467008]]
+        true_xbins = np.array([-3.79679950e+00,  -3.41718685e+00,  -3.03757420e+00,
+                               -2.65796156e+00,  -2.27834891e+00,  -1.89873626e+00,
+                               -1.51912362e+00,  -1.13951097e+00,  -7.59898323e-01,
+                               -3.80285676e-01,  -6.73029298e-04,   3.78939617e-01,
+                               7.58552264e-01,   1.13816491e+00,   1.51777756e+00,
+                               1.89739020e+00,   2.27700285e+00,   2.65661550e+00,
+                               3.03622814e+00,   3.41584079e+00,   3.79545344e+00])
+
+        assert_true(np.allclose(count, true_count))
+        assert_true(np.allclose(med, true_med))
+        assert_true(np.allclose(conf[:, 0], true_conf[0]))
+        assert_true(np.allclose(conf[:, 1], true_conf[1]))
+        assert_true(np.allclose(xbins, true_xbins))
+
+        # plt.clf(); plt.scatter(xx, yy, color='b', alpha=0.5)
+        # zcode.plot.plot_core.plotHistLine(plt.gca(), xbins, med, c='k', lw=2.0)
+        # zcode.plot.plot_core.plotHistLine(plt.gca(), xbins, conf[:, 0], c='green', lw=2.0)
+        # zcode.plot.plot_core.plotHistLine(plt.gca(), xbins, conf[:, 1], c='green', lw=2.0)
+        # zcode.plot.plot_core.plotHistLine(plt.gca().twinx(), xbins, med, c='k', lw=2.0)
 
     def test_smooth(self):
         r2 = self.r2